+++ lib/elixir_script/lib/js.ex
@@ -41,14 +41,17 @@ defmodule ElixirScript.JS do
   @doc """
   Mutates an existing JavaScript object.
   ex:
-    ElixirScript.JS.mutate elem, %{"width" => 100}
-  """
-  defmacro update(object, map)
-
-  @doc """
-  Mutates an existing JavaScript object.
-  ex:
     ElixirScript.JS.mutate elem, "width", 100
   """
-  defmacro update(object, key, value)
+  defmacro mutate(object, key, value)
+
+
+  @doc """
+  Takes the given map and returns an object
+  Throws an error if any key is not a
+  number, binary, or atom
+  ex:
+    ElixirScript.JS.map_to_object(%{my: "map"})
+  """
+  defmacro map_to_object(object)
 end