+++ src/javascript/tests/core/erlang_compat/maps_spec.js
@@ -2,7 +2,7 @@ import test from 'ava';
 import Core from '../../../lib/core';
 
 test('find', t => {
-  let myMap = {};
+  let myMap = new Map();
   let result = Core.maps.find('t', myMap);
   t.is(result, Symbol.for('error'));
 
@@ -10,13 +10,13 @@ test('find', t => {
   result = Core.maps.find('t', myMap);
   t.deepEqual(result.values, [Symbol.for('badmap'), myMap]);
 
-  myMap = { t: 'b' };
+  myMap = new Map([['t', 'b']]);
   result = Core.maps.find('t', myMap);
   t.deepEqual(result.values, [Symbol.for('ok'), 'b']);
 });
 
 test('fold', t => {
-  const myMap = { a: 1, b: 2 };
+  const myMap = new Map([['a', 1], ['b', 2]]);
   const result = Core.maps.fold((k, v, acc) => acc + v, 0, myMap);
   t.is(result, 3);
 });